# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basis Social API",
    "version": "1.0.0-alpha",
    "description": "Stable REST and world-server integration contract for BasisVR. ActivityPub endpoints are intentionally outside this client API contract."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/v1/users/{id}/block": {"post": {"tags":["Relationships"],"operationId":"blockUser","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"Blocked"},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/users/{id}/unblock": {"post": {"tags":["Relationships"],"operationId":"unblockUser","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"Unblocked"},"default":{"$ref":"#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type":"http","scheme":"bearer","bearerFormat":"JWT"},
      "serviceToken": {"type":"apiKey","in":"header","name":"X-Basis-Service-Token"}
    },
    "parameters": {
      "ID": {"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
    },
    "responses": {
      "Error": {"description":"Error response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}
    },
...
    "schemas": {
      "Error": {"type":"object","required":["error"],"properties":{"error":{"type":"object","required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"}}}}},
...
    }
  }
}